Improved NERDTree and Taglist setups.
A few lines were included in vimrc so it can now behave more like an IDE.
NERDTree and Taglist are now displayed right at Vim's startup, each one in the
first and last collumns of the window, respectively.

The NERDTree key mapping has also been improved. <F3> is acting as a toggle
button, closing and opening NERDTree as it is pressed.

A new mapping has also been included for Taglist. Just as in NERDTree's
case, Taglist's key (<F4>) also acts like a toggle button to close and
open its window.